+++ src/refs/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./frontmatter";
 export * from "./router";
+export * from "./loader";